+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 * Handle `srcset` entries without a width or pixel density descriptor.
 * Embed subresources recursively when calling `WebResource.to_data_uri()` on an archive's main resource.
+* Don't escape HTML entities in a `<script>` or `<style>` block.
 
 ## [0.3.0] - 2021-07-18
 ### Added